+++ vlaanderen/CHANGELOG.md
@@ -9,6 +9,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [1.0.2-SNAPSHOT]
 - Adding download link for RDF
 - Enabled 'pin as favorite' functionality
+- Adding VL/INSPIRE conformity keywords upon validation
 - [core] high availability fix, allow configuration of separate html cache directory / [pr](https://agiv.visualstudio.com/Metadata/_git/MetadataGeonetwork/pullrequest/24976)
 
 ## [1.0.1] - 2023-06-15